Replace the original Examiner with new VDOM variant
