+++ bokehjs/src/lib/models/ui/index.ts
@@ -4,7 +4,6 @@ export * from "./menus"
 export {Dialog} from "./dialog"
 export {Drawer} from "./drawer"
 export {Examiner} from "./examiner"
-export {ExaminerVDOM} from "./examiner_vdom"
 export {Notifications} from "./notifications"
 export {Panel} from "./panel"
 export {Pane} from "./pane"
